Search for end of frame from front to back
MLLP senders do not necessarily wait for an ACK to send more MLLP
messages, meaning that the semantic of searching for the end of the
message at the end of the current TCP socket buffer allows a "frame"
to contain more than one message so long as the start and end bytes
of the stack of messages are correct.

Rewire the search loop to search front-to-back, which will mean
iterating all bytes on the socket in cases where sender requires
an ACK before sending more messages, but will work for asynchronous
transports which do not require or simply ignore the ACK. This
ensures semantically correct parsing, at the expense of more stream
processing in the iterator.